requirements.txt
https://github.com/django/django/archive/refs/tags/4.1.tar.gz
psycopg2
python-dotenv==0.21.0
whitenoise==6.2.0
gunicorn==20.1.0
django-extensions
